requirements.txt
# Copy of backend requirements for Vercel deployment
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0
aioredis>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
httpx>=0.25.0
requests>=2.31.0
slowapi>=0.1.8
structlog>=23.2.0
asyncio-mqtt>=0.16.1
celery>=5.3.0
langchain>=0.1.0
openai>=1.3.0
anthropic>=0.7.0